replaced unsigned int by size_t in string positions

+++ architecture/faustqt.h
@@ -797,8 +797,8 @@ class dbHorizontalBargraph : public dbBargraph
  */
 static string rmWhiteSpaces(const string& s)
 {
-    unsigned int i = s.find_first_not_of(" \t");
-    unsigned int j = s.find_last_not_of(" \t");
+    size_t i = s.find_first_not_of(" \t");
+    size_t j = s.find_last_not_of(" \t");
 
 	if ( (i != string::npos) & (j != string::npos) ) {
 		return s.substr(i, 1+j-i);
